Addressing Theo's comments
