+++ src/enrichment/models/base.rs
@@ -60,25 +60,6 @@ pub trait Model {
         Ok(triplets)
     }
 
-    fn get_triplet_for_cider(
-        &self,
-        alerts: &[Document],
-    ) -> Result<Array<f32, Dim<[usize; 4]>>, ModelError> {
-        let mut triplets = Array::zeros((alerts.len(), 3, 49, 49));
-        for i in 0..alerts.len() {
-            let (cutout_science, cutout_template, cutout_difference) = prepare_triplet(&alerts[i])?;
-            for (j, cutout) in [cutout_science, cutout_template, cutout_difference]
-                .iter()
-                .enumerate()
-            {
-                let mut slice = triplets.slice_mut(ndarray::s![i, j, .., ..]);
-                let full_array = Array::from_shape_vec((63, 63), cutout.to_vec())?;
-                let cutout_array = full_array.slice(ndarray::s![7..56, 7..56]).to_owned();
-                slice.assign(&cutout_array);
-            }
-        }
-        Ok(triplets)
-    }
     fn predict(
         &mut self,
         metadata_features: &Array<f32, Dim<[usize; 2]>>,
